Tweaks poi-preset-karma (#265)
* made changes :)

* forgot to commit this

* new: Exposes reporters in preset options

`reporters` default to `[mocha]`. If other reporters are set in the
preset options they replace the default. In either case `coverage`
will be added based on the coverage setting (`coverage: true` or
`--coverage` on the command line).

If you want to role your own coverage reporting the `karma:` poi
options completely override the preset settings.