Support customized name for 'vendor', 'manifest' and 'client'
Now, If you are bundling you app that depends on third-party libraries by poi, you can customize the names of 'vendor', 'manifest' and 'client' file.

And '[name]' is the shortcut of the name in 'package.json'. if not found name in package.json, fallbacks to current folder name in kebab case.